+++ CHANGELOG.md
@@ -8,14 +8,25 @@ vNext
 -
 -
 -
+- Added lifted conditional `nn.cond`
 -
+-
+-
+-
+-
+-
+-
+-
+
+0.4.1
+-----
+
+New features:
+- Added locally-connected (unshared CNN) layer `flax.linen.ConvLocal`.
 - Improved seq2seq example: Factored our model and input pipeline code.
 - Added Optax update guide and deprecated `flax.optim`.
 - Added `sep` argument to `flax.traverse_util.flatten_dict()`.
 - Implemented Sequential module, in `flax.linen.combinators`.
--
-- Added locally-connected (unshared CNN) layer `flax.linen.ConvLocal`.
--
 
 0.4.0
 ------